+++ app/features/instruction-program-mango/ui/AddSpotMarketDetailsCard.tsx
@@ -1,7 +1,6 @@
-import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
-
-import { InstructionCard } from '../InstructionCard';
+import { InstructionCard } from '@components/instruction/InstructionCard';
 import { AddSpotMarket, spotMarketFromIndex } from '@explorer/decoder-mango';
+import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
 
 export function AddSpotMarketDetailsCard(props: {
     ix: TransactionInstruction;
