+++ app/features/instruction-program-mango/ui/ChangePerpMarketParamsDetailsCard.tsx
@@ -1,10 +1,13 @@
+import { InstructionCard } from '@components/instruction/InstructionCard';
+import {
+    ChangePerpMarketParams,
+    getPerpMarketFromInstruction,
+    getPerpMarketFromPerpMarketConfig,
+} from '@explorer/decoder-mango';
 import { useCluster } from '@providers/cluster';
 import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
 import moment from 'moment';
 import { useEffect, useState } from 'react';
-
-import { InstructionCard } from '../InstructionCard';
-import { ChangePerpMarketParams, getPerpMarketFromInstruction, getPerpMarketFromPerpMarketConfig } from '@explorer/decoder-mango';
 
 export function ChangePerpMarketParamsDetailsCard(props: {
     ix: TransactionInstruction;
