+++ app/features/instruction-program-mango/ui/GenericMngoAccountDetailsCard.tsx
@@ -1,7 +1,6 @@
 import { Address } from '@components/common/Address';
+import { InstructionCard } from '@components/instruction/InstructionCard';
 import { SignatureResult, TransactionInstruction } from '@solana/web3.js';
-
-import { InstructionCard } from '../InstructionCard';
 
 export function GenericMngoAccountDetailsCard(props: {
     ix: TransactionInstruction;
